auto-email-responder/requirements.txt
google-api-python-client
google-auth-oauthlib
google-auth
sentence-transformers
faiss-cpu
openai
cachetools
beautifulsoup4
numpy
schedule
requests
langchain
langchain-openai
langchain-google-community
langchain-community
jq
langchainhub
python-dotenv